Adding color button widget and updating plugins.
This commit adds a subclass of QPushButton called ColorButton that
implements a widget for displaying and selecting colors.  We've been
doing this manually everywhere with duplicated code.  This is a simple
abstraction but allows us to elminate a lot of duplication, especially
in plugins that have multiple color selections.

Conflicts:
	mapviz_plugins/include/mapviz_plugins/gps_plugin.h
	mapviz_plugins/src/gps_config.ui
	mapviz_plugins/src/gps_plugin.cpp