Update map canvas at a fixed rate.
This update adds a timer to the map canvas to repaint at a fixed rate.
The default rate is 50 Hz, but there is a method to change it (not
exposed to the UI at the moment).  50Hz was chosen because it is fast
enough to give smooth animations and we almost always are running
mapviz with at least one plugin triggering updates from a 50Hz topic.