Fixing #317
First, the model view matrix needs to be saved and restored around
QPainter operations because Qt clears several GL variables.  Also, the
image plugin needed to explicitly call glMatrixMode(GL_PROJECTION);
it does a few operations on the projection matrix and was just assuming
that was the current matrix mode.  Also, I added a function that plugins
need to override if they want to do QPainter operations; this will
eliminate unnecessary overhead for plugins that do not.