Migrate route plugin to use swri_route_util
This change migrates the mapviz route plugin to use swri_route_util to
get consistent behavior with route transforms and route position
interpolation.  As part of this change, the route is now transformed
with each draw so that it will correctly move around if the transform
between the fixed frame and the route frame is not constant.