Migrate route plugin to use swri_route_util
This change migrates the mapviz route plugin to use swri_route_util to
get consistent behavior with route transforms and route position
interpolation.  As part of this change, the route is now transformed
with each draw so that it will correctly move around if the transform
between the fixed frame and the route frame is not constant.

 # Please enter the commit message for your changes. Lines
starting # with '#' will be ignored, and an empty message aborts the
commit.  # On branch migrate-to-swri-route-util # Changes to be
committed: # modified: CMakeLists.txt # modified:
include/mapviz_plugins/route_plugin.h # modified: package.xml #
modified: src/route_plugin.cpp #