Add option to not scale arrows with zoom level
This adds a checkbox to all of the plugins that can draw a series of
coordinates as arrows; i. e., the GPS, NavSatFix, Odometry, and TF Frame
plugins.  This checkbox will control whether the arrows are drawn at a fixed
size regardless of zoom level or whether they are scaled with the zoom level.

Resolves #414