Rewrite tile_map loading to be more reliable
This changes how the tile_map plugin handles making network requires for tiles.
It will now:
- Use thread conditions to prompt loading rather than spinning
- Prioritize loading tiles in the visible area
- Retry a failed network request up to 5 times
- Not discard tile requests if there are more than 100 in the queue

This changes should significantly reduce (if not completely eliminate) the
number of tiles that fail to load and hopefully make tiles within the visible
area appear faster when there are many in the queue.

Fixes #342 and #421.