Move video recording into its own thread
This creates a worker that handles the image conversion and file writing for
video recording and places it in its own thread.  Video recording is fairly
CPU intensive, so when video recording is enabled, this should provide for
a smoother GUI as well as a more consistent frame rate in the video.