Fix issue with loading Bing map tiles
The Bing API requires making an initial authorization call before it can
generate URLs for requesting tiles.  A side effect of this was that if
the tile_map plugin tried to request any times before that initial call
completed, those tiles would fail to load, and they would *never* load
properly after that.

This adds a mechanism for the tile source to indicate whether it is
ready to load tiles or not, and the tile_map plugin will wait until it
is ready before requesting images, so Bing maps should always load
properly on startup.